requirements.txt
requests==2.32.3
python-dotenv==1.0.1
SQLAlchemy==2.0.31
psycopg==3.1.19
psycopg2-binary==2.9.9
alembic==1.13.2
fastapi==0.111.0
dependency-injector==4.41.0
uvicorn==0.30.1
gunicorn==22.0.0
asyncpg==0.29.0
greenlet==3.0.3
loguru==0.7.2